Admin - Move internal EC2 models (#5090)
