+++ moto/ec2/models/amis.py
