+++ moto/ec2/models/availability_zones_and_regions.py
