+++ moto/ec2/models/elastic_ip_addresses.py
