+++ moto/ec2/models/iam_instance_profile.py
