+++ moto/ec2/models/instance_types.py
