+++ moto/ec2/models/internet_gateways.py
