+++ moto/ec2/models/key_pairs.py
