+++ moto/ec2/models/launch_templates.py
