+++ moto/ec2/models/nat_gateways.py
