+++ moto/ec2/models/transit_gateway.py
