+++ moto/ec2/models/transit_gateway_attachments.py
