+++ moto/ec2/models/vpc_service_configuration.py
