+++ moto/ec2/models/vpn_connections.py
