+++ moto/ec2/models/vpn_gateway.py
